Finished Acceptance Tests (almost)
Added the database and the acceptance tests to satisfy the criteria. Spent a lot of time toying with the various ways to load csv into SQL. 

Still missing an acceptance test to make sure the update / add return false. Need to stub out the repository for those tests but want to take a crack at React before the weekend is over.