ScoreFormatter + Rework due to misunderstanding how Tennis works
Added a set score formatter, which processes the set data into the desired format.

Due to my inital lack of knowledge regarding tennis and its terminology I misunderstood the requirments. It seems that each line of input is not actually stand-alone but must consider the previous line.

There was a re-work of the scores class to reflect this, definitely some more work deeper in the Api that needs re-working to pass this information to it (set calculator, as it is going to need to tell the set score formatter the initial servers - not sure whether this will handle complex cases where servers change mid input but there you go).

Next step is specflow to help with that process.

Also added an extension method to go from internal scores to tennis scores to handle deuce sets or whatever they may actually be called.